pyproject.toml
[project]
name = "arke-terminal"
version = "0.1.192"
description = "Local-first RAG for legal teams."
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Pavel Dolgoter", email = "mikepromogratus@proton.me" },
]
readme = "README.md"
keywords = ["rag", "search", "embeddings", "legal", "local"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "python-dotenv>=1.0",
    "numpy>=2.0",
    "click>=8.1",
    "llama-cpp-python>=0.3",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "extract-msg>=0.48",
    "msal>=1.36.0",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.urls]
Repository = "https://github.com/padolgot/arke-terminal"

[project.scripts]
arke-server  = "arke.server.main:run"
arke-gateway = "arke.clients.socket:run"
arke         = "arke.clients.cli:main"
arke-eval    = "arke.eval.sweep:run"
arke-mail    = "arke.clients.email:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["arke"]
